ImGui version is stable enough to be merged in the main branch (#1)
* ImGUI implementation

* Rough something

* Clean up unneeded files

* Linux fixes and mapper fix (AGAIN)

* some stuff. needs a lot of refactoring.

because I am learning ImGui

* more stuff

* Bus reset support added.

Trying to reset when the game CastleVania the game didn't reset. But the CPU did. ColdReset now will ask all bus_device(s) to clear their contents by calling the busreset() function on the bus object.

* Update DenverEMU.cpp

* Update mapper_001.cpp

* trying to fix cart.cpp linux crash MMC1

* Update DenverEMU.cpp

* Refactoring, cleaning up.

* Update ppu.cpp

* Finding bugs - n overflow to eight or higher ppu

* state fixes. must look into it for real behaiviour

* test on real hardware. starting to think it is VMware

* Finally fixed the cart code

* Trying to get performance.. Function calls slow it down.

* Update mapper_001.cpp

* Moved GUI code to seperate source files

* Gui enhancements open dialog implemented.

* Update dengui_main.cpp

* Beginning implementation of dialogs

* Optimizations, combining all 2a03 related devices to a singular device.

also updated bus logic. When a device is attached it will call a function registering the bus on which the device is attached to the device.

* stupid vc 2017 allows for stuff such as forgetting #include <string>

* Modified code to use callbacks.

Windoiws builds got enormous boost. 180fps -> 590fps. Setting sync states between ppu<->cpu has no effect. (always maximum speed!)

* Fixes: APU, Package, Emulation speed

Emulation drasticly improved. PPU <-> CPU sync on cycle is as fast as 200 cycles per sync.
this makes the sync function utterly useless and will be deprecated for the next commit as it does nothing now.
This enables permanent full accuracy when the ppu code is finished.

---------

+++ video/nesvideo.h
@@ -27,7 +27,7 @@ static const std::uint8_t ntscpalette[] = {
 
 class nesvideo {
 private:
-	std::uint16_t* displaybuffer;	// 16bit
+	std::uint32_t* displaybuffer;	// 16bit
 public:
 	nesvideo();
 	~nesvideo();
